requirements.txt
Flask==2.3.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1 